requirements.txt
crewai
youtube-transcript-api
python-dotenv